Major Code Refactoring
- Seperated type conversions from MarshalType into new class TypeConverter
- Seperated read/write operations from IO into new classes Literate & Manager
- Added some missing struct constraints for generic types
- Added memory allocate, free & copy functions to Manager
- Implemented managed memory allocation
- Change some datatypes from IntPtr to int/long for ease of use
- Renamed Protection.cs to ManagedProtection.cs
- Renamed and optimized a bunch of other stuff
- Fix build for x86

+++ BlueMagic/Native/Structures.cs
@@ -9,7 +9,7 @@ namespace BlueMagic.Native
         public IntPtr BaseAddress;
         public IntPtr AllocationBase;
         public MemoryProtectionType AllocationProtect;
-        public IntPtr RegionSize;
+        public int RegionSize;
         public MemoryAllocationState State;
         public MemoryProtectionType Protect;
         public MemoryAllocationType Type;
